refactor(Sample.Core): nest GitHubSampleCodeFetcher + Fetch* companions under Sample.Core; rename file
Continues Sample.Core type-level coverage from #357. The GitHub fetcher
class + its four companion types in Core/GitHubSampleCodeFetcher.swift
now live under Sample.Core, with the file renamed to match.

SampleCodeDownloader.swift is the other Sample.Core type-level pending
from #357/#369 — it ships in its own PR because wrapping the actor in
'extension Sample.Core' shifts the @MainActor isolation context for
file-scope helper classes (AuthFlowCoordinator etc.) and trips several
Swift 6 concurrency errors that need targeted fixes. Doing it alongside
GitHubFetcher would have made the review surface murky.

## Renames

| Before | After |
|---|---|
| GitHubSampleCodeFetcher (public final class) | Sample.Core.GitHubFetcher |
| FetchStatistics                              | Sample.Core.FetchStatistics |
| FetchAction                                  | Sample.Core.FetchAction |
| FetchProgress                                | Sample.Core.FetchProgress |
| GitHubFetcherError                           | Sample.Core.GitHubFetcherError |

The four companions kept their names — they were already
sample-fetch-scoped, no prefix to drop.

## File rename

| Before | After |
|---|---|
| Core/GitHubSampleCodeFetcher.swift | Core/Sample.Core+GitHubFetcher.swift |
                                       (dominant: GitHubFetcher; carries
                                       FetchStatistics, FetchAction,
                                       FetchProgress, GitHubFetcherError
                                       companions)

## Mechanics

- Each top-level public type wrapped in 'extension Sample.Core { ... }'.
- Internal references to GitHubSampleCodeFetcher within the file
  rewritten to GitHubFetcher (sibling refs work from inside the
  extension).
- 1 caller swept: CLI/Commands/Command+Fetch.swift.

## Verification

- xcrun swift build clean.
- xcrun swift test: 1300/1300 passing.

Part of the namespacing sweep tracked in #183. Sample-related history:
#356 (foundation + Sample.Cleanup.Cleaner), #357 (Sample.Core
initial — Catalog, Entry, Statistics, Progress, Project),
#358 (Sample.Services + Sample.Search), #360 (Sample.Indexer +
Sample.Atom), #361 (Sample.Index), #362 (Sample.Format.*),
#369 (cross-cutting file renames). Remaining Sample.Core work:
SampleCodeDownloader (its own PR; @MainActor handling).
